Rename ファイルフォーカス機能の名前空間の修正
